pyproject.toml
[project]
name = "writing-assistant-pro"
version = "2.0.0"
description = "Inspired by writting tool app"
authors = [
    { name = "3C0D", email = "author@example.com" }
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
dependencies = [
    "flet>=0.24.0", # Flutter-based UI framework for Python (native desktop)
    "babel", # For translation management
    "keyboard", # For global hotkey management
    "loguru", # Modern logging library with colored output
    "flet-desktop>=0.28.3",
    "pystray>=0.19.5",
    "pillow>=12.0.0",
    "requests>=2.32.0", # For GitHub API calls (update checking)
    "packaging>=24.0", # For version comparison (update checking)
    "pyperclip>=1.9.0", # For clipboard text handling
    "pynput>=1.7.6", # For keyboard simulation (Ctrl+C to capture selection)
]

[dependency-groups]
dev = [
    "pytest>=9.0.1", # Testing framework
    "ruff", # Fast Python linter and formatter
    "pyright", # Type checker
    "pre-commit", # For managing pre-commit hooks
    "watchdog>=6.0.0", # File system monitoring, used to detect css changes
    "pyinstaller>=6.0.0", # Package application into executable
    "commitizen>=4.10.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py313"  # Target Python
cache-dir = ".ruff_cache"  # For caching Ruff's analysis results
exclude = ["build", "dist", "__pycache__", "myvenv", ".uv", ".pytest_cache", ".mypy_cache", ".venv", ".git", ".ruff_cache"]  # Exclude from linting

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "C4", "B"]  # Enabled rule categories: pycodestyle errors, warnings, pyflakes, isort, pyupgrade, flake8-comprehensions, flake8-bugbear
ignore = []  # List of rules to ignore

[tool.ruff.lint.isort]
known-first-party = ["src"]  # Treat 'src' as first-party imports for import sorting

[tool.pyright]
extraPaths = ["src"]
pythonVersion = "3.13"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.0.0"
version_provider = "pep621"
version_files = [
    "src/version.py:__version__",
    "pyproject.toml:project.version"
]
tag_format = "v$version"
update_changelog_on_bump = true
bump_message = "bump: version $current_version → $new_version"
